exception handiling added

+++ src/main/java/com/rest/service/BookService.java
@@ -1,6 +1,8 @@
 package com.rest.service;
 
 import java.util.List;
+
+import com.rest.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.rest.model.Book;
@@ -11,13 +13,16 @@ public class BookService {
 
 	@Autowired
 	BookRepository bookRepository;
+	Book book;
 
 	public List<Book> getAllBooks() {
 		return bookRepository.findAll();
 	}
 
 	public Book getBookById(Integer id) {
-		return bookRepository.findOne(id);
+		book= bookRepository.findOne(id);
+		if(book==null){throw new ResourceNotFoundException(id,"book not found");}
+		return book;
 	}
 
 	public void saveBook(Book book) {
@@ -25,6 +30,8 @@ public class BookService {
 	}
 
 	public void deleteBook(Integer id) {
+		book= bookRepository.findOne(id);
+		if(book==null){throw new ResourceNotFoundException(id,"book not found");}
 		bookRepository.delete(id);
 	}
 }
